+++ readme.xcodeproj/xcuserdata/adamarmstrong.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>ReadMe.xcscheme_^#shared#^_</key>
+		<key>readme.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
