--- readme/Assets.xcassets/BookTexture.imageset/Contents.json
@@ -1,32 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "LightTexture.jpg",
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "light"
-        }
-      ],
-      "filename" : "LightTexture-1.jpg",
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "DarkTextureBlack.jpg",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
